+++ lib/src/constants/app_constants.dart
@@ -4,4 +4,10 @@
 class KartalAppConstants {
   static const String IPAD_TYPE = 'ipad';
   static const String WHATS_APP_PREFIX = 'whatsapp://send?text=';
+
+  static const String GOOGLE_MAPS_WEB_LINK =
+      'https://www.google.com/maps/search/?api=1&query=';
+
+  static const String GOOGLE_MAPS_DEEP_URL = 'geo:0,0?q=';
+  static const String APPLE_MAPS_DEEP_URL = 'maps://?q=';
 }